Move class-related tests to a directory
