+++ test/unit/classes/accessors.spec.ts
@@ -1,4 +1,4 @@
-import * as util from "../util";
+import * as util from "../../util";
 
 test("get accessor", () => {
     util.testFunction`
